requirements.txt
flask
tensorflow
tensorflow-hub
pillow
numpy
